source/dataagent-core/pyproject.toml
[project]
name = "dataagent-core"
version = "0.1.0"
description = "Core library for DataAgent - Agent creation, execution, and event system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.8",
    "langchain>=0.3.7",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "requests>=2.31.0",
    "tavily-python>=0.3.0",
    "markdownify>=0.13.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
sandbox = [
    "modal>=0.65.0",
    "daytona>=0.113.0",
    "runloop-api-client>=0.69.0",
]
mysql = [
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
]
mcp = [
    "langchain-mcp-adapters>=0.0.1",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dataagent_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["dataagent_core"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
